+++ includes/ziti/ziti_model.h
@@ -47,14 +47,34 @@ XX(encryption, bool, none, encryptionRequired, __VA_ARGS__) \
 XX(perm_flags, int, none, NULL, __VA_ARGS__) \
 XX(config, json, map, config, __VA_ARGS__)
 
-#define ZITI_INTERCEPT_MODEL(XX, ...) \
+#define ZITI_CLIENT_CFG_V1_MODEL(XX, ...) \
 XX(hostname, string, none, hostname, __VA_ARGS__) \
 XX(port, int, none, port, __VA_ARGS__)
+
+#define ZITI_PORT_RANGE_MODEL(XX, ...) \
+XX(low, int, none, low, __VA_ARGS__) \
+XX(high, int, none, high, __VA_ARGS__)
+
+#define ZITI_INTERCEPT_CFG_V1_MODEL(XX, ...) \
+XX(protocols, string, array, protocols, __VA_ARGS__) \
+XX(addresses, string, array, addresses, __VA_ARGS__) \
+XX(port_ranges, ziti_port_range, array, portRanges, __VA_ARGS__) \
+XX(dial_options, tag, map, dialOptions, __VA_ARGS__) \
+XX(source_ip, string, none, sourceIp, __VA_ARGS__)
 
 #define ZITI_SERVER_CFG_V1_MODEL(XX, ...) \
 XX(protocol, string, none, protocol, __VA_ARGS__) \
 XX(hostname, string, none, hostname, __VA_ARGS__) \
 XX(port, int, none, port, __VA_ARGS__)
+
+#define ZITI_HOST_CFG_V1_MODEL(XX, ...) \
+XX(protocol, string, none, protocol, __VA_ARGS__) \
+XX(dial_intercepted_protocol, bool, none, dialInterceptedProtocol, __VA_ARGS__) \
+XX(address, string, none, address, __VA_ARGS__) \
+XX(dial_intercepted_address, bool, none, dialInterceptedAddress, __VA_ARGS__) \
+XX(port, int, none, port, __VA_ARGS__) \
+XX(dial_intercepted_port, bool, none, dialInterceptedPort, __VA_ARGS__) \
+XX(listen_options, tag, map, listenOptions, __VA_ARGS__)
 
 #ifdef __cplusplus
 extern "C" {
@@ -76,9 +96,15 @@ DECLARE_MODEL(ziti_identity, ZITI_IDENTITY_MODEL)
 
 DECLARE_MODEL(ziti_service, ZITI_SERVICE_MODEL)
 
-DECLARE_MODEL(ziti_intercept, ZITI_INTERCEPT_MODEL)
+DECLARE_MODEL(ziti_client_cfg_v1, ZITI_CLIENT_CFG_V1_MODEL)
+
+DECLARE_MODEL(ziti_port_range, ZITI_PORT_RANGE_MODEL)
+
+DECLARE_MODEL(ziti_intercept_cfg_v1, ZITI_INTERCEPT_CFG_V1_MODEL)
 
 DECLARE_MODEL(ziti_server_cfg_v1, ZITI_SERVER_CFG_V1_MODEL)
+
+DECLARE_MODEL(ziti_host_cfg_v1, ZITI_HOST_CFG_V1_MODEL)
 
 ZITI_FUNC const char *ziti_service_get_raw_config(ziti_service *service, const char *cfg_type);